+++ vetiver/handlers/torch.py
@@ -1,26 +1,31 @@
 from vetiver.handlers.base import VetiverHandler
-from ..ptype import vetiver_create_ptype
 from ..meta import vetiver_meta
+from ..ptype import vetiver_create_ptype
+import numpy as np
 
-import pandas as pd
-import sklearn
+torch_exists = True
+try:
+    import torch
+except ImportError:
+    torch_exists = False
 
-class SKLearnHandler(VetiverHandler):
-    """Handler class for creating VetiverModels with sklearn.
+
+class TorchHandler(VetiverHandler):
+    """Handler class for creating VetiverModels with torch.
 
     Parameters
     ----------
-    model : sklearn.base.BaseEstimator
-        a trained sklearn model
+    model : nn.Module
+        a trained torch model
     """
-
+    base_class = torch.nn.Module
     def __init__(self, model, ptype_data):
         super().__init__(model, ptype_data)
 
     def create_description(self):
-        """Create description for sklearn model
+        """Create description for torch model
         """
-        desc = f"Scikit-learn {self.model.__class__} model"
+        desc = f"Pytorch model of type {type(self.model)}"
         return desc
 
     def vetiver_create_meta(
@@ -29,8 +34,9 @@ class SKLearnHandler(VetiverHandler):
         url: str = None,
         required_pkgs: list = [],
     ):
-        """Create metadata for sklearn model"""
-        required_pkgs = required_pkgs + ["scikit-learn"]
+        """Create metadata for torch model
+        """
+        required_pkgs = required_pkgs + ["torch"]
         meta = vetiver_meta(user, version, url, required_pkgs)
 
         return meta
@@ -49,6 +55,7 @@ class SKLearnHandler(VetiverHandler):
             Zero-row DataFrame for storing data types
         """
         ptype = vetiver_create_ptype(self.ptype_data)
+
         return ptype
 
     def handler_startup():
@@ -58,7 +65,6 @@ class SKLearnHandler(VetiverHandler):
         function for tasks like loading packages.
         """
         ...
-
 
     def handler_predict(self, input_data, check_ptype):
         """Generates method for /predict endpoint in VetiverAPI
@@ -77,18 +83,17 @@ class SKLearnHandler(VetiverHandler):
         prediction
             Prediction from model
         """
+        if torch_exists:
+            if check_ptype == True:
+                input_data = np.array(input_data, dtype=np.array(self.ptype_data).dtype)
+                prediction = self.model(torch.from_numpy(input_data))
+            
+            # do not check ptype
+            else:    
+                input_data = torch.tensor(input_data)
+                prediction = self.model(input_data)
 
-        if check_ptype == True:
-            if isinstance(input_data, pd.DataFrame):
-                prediction = self.model.predict(input_data)
-            else:
-               prediction = self.model.predict([input_data])
-
-        # do not check ptype
         else:
-            if not isinstance(input_data, list):
-                input_data = [input_data.split(",")]  # user delimiter ?
-
-            prediction = self.model.predict(input_data)
+            raise ImportError("Cannot import `torch`.")
 
         return prediction
